+++ include/caffe/layers/permute_layer.hpp
@@ -17,7 +17,7 @@ namespace caffe {
 
 // The main function which does the permute.
 template <typename Dtype>
-void Permute(const int count, Dtype* bottom_data, const bool forward,
+void Permute(const long count, Dtype* bottom_data, const bool forward,
     const int* permute_order, const int* old_steps, const int* new_steps,
     const int num_axes, Dtype* top_data);
