+++ src/caffe/layers/dropout_layer.cpp
@@ -32,7 +32,7 @@ void DropoutLayer<Ftype, Btype>::Forward_cpu(const vector<Blob*>& bottom,
   const Ftype* bottom_data = bottom[0]->cpu_data<Ftype>();
   Ftype* top_data = top[0]->mutable_cpu_data<Ftype>();
   unsigned int* mask = rand_vec_.mutable_cpu_data();
-  const int count = bottom[0]->count();
+  const long count = bottom[0]->count();
   if (this->phase_ == TRAIN) {
     // Create random numbers
     caffe_rng_bernoulli<float>(count, 1. - threshold_, mask);
@@ -53,7 +53,7 @@ void DropoutLayer<Ftype, Btype>::Backward_cpu(const vector<Blob*>& top,
     Btype* bottom_diff = bottom[0]->mutable_cpu_diff<Btype>();
     if (this->phase_ == TRAIN) {
       const unsigned int* mask = rand_vec_.cpu_data();
-      const int count = bottom[0]->count();
+      const long count = bottom[0]->count();
       for (int i = 0; i < count; ++i) {
         bottom_diff[i] = top_diff[i] * mask[i] * scale_;
       }
