+++ src/caffe/layers/euclidean_loss_layer.cpp
@@ -17,7 +17,7 @@ void EuclideanLossLayer<Ftype, Btype>::Reshape(
 template <typename Ftype, typename Btype>
 void EuclideanLossLayer<Ftype, Btype>::Forward_cpu(const vector<Blob*>& bottom,
     const vector<Blob*>& top) {
-  int count = bottom[0]->count();
+  long count = bottom[0]->count();
   caffe_sub<Ftype>(
       count,
       bottom[0]->cpu_data<Ftype>(),
