+++ src/caffe/layers/euclidean_loss_layer.cu
@@ -8,7 +8,7 @@ namespace caffe {
 template <typename Ftype, typename Btype>
 void EuclideanLossLayer<Ftype, Btype>::Forward_gpu(const vector<Blob*>& bottom,
     const vector<Blob*>& top) {
-  int count = bottom[0]->count();
+  long count = bottom[0]->count();
   caffe_gpu_sub<Ftype>(
       count,
       bottom[0]->gpu_data<Ftype>(),
